refactor

+++ shoppingcart/src/main/java/org/example/Item.java
@@ -1,4 +1,6 @@
 package org.example;
 
-public record Item(String barcodeNum, String name, double price) {
+import java.math.BigDecimal;
+
+public record Item(String barcodeNum, String name, BigDecimal price) {
 }
